Revise testGetForecast to not make live API calls
While attempting to run the tests, testGetForecast consistently failed,
with the exception's error message saying that the BOM API had detected
the use of automation software during the request. Given that, and that
it may well not be necessary to actually make a live request, this
change updates the test to use a mock response with one of the fixture
files instead. It also improves the performance of the test.